syscall: carry full u64 args through dispatch_inner, widen pointer-taking handlers
The 64-bit-vamap migration moves user space into the upper canonical-low
half (above 4 GiB), so any syscall arg that holds a user pointer must be
preserved as u64 across the dispatch path. The previous boundary
truncated args to u32 right before dispatch_inner, which crashed every
pointer-bearing handler the moment a user pointer landed above 4 GiB
(seen first as the compositor SIGSEGV on SYS_MAP_FRAMEBUFFER, then on
SYS_TIME, etc.).

- dispatch_inner now takes (u32, u64, u64, u64, u64, u64); both x86
  syscall_dispatch_64 and arm64_syscall_dispatch pass args untruncated.
- ~80 pub fn sys_*(_ptr: u32, ...) handlers in 13 files widened to u64;
  internals updated where arithmetic/casts assumed 32-bit (e.g.
  buf_ptr.wrapping_add(total as u64) in sys_write).
- sys_thread_create / sys_sigaction widened too — entry_rip, user_rsp,
  handler_addr are user addresses, not scalars.
- sys_map_framebuffer routed through the 64-bit special-case path in
  syscall_dispatch_64 alongside MMAP/MUNMAP/AVM, since it takes a u64
  out-pointer (FbMapInfo).
- helpers (read_user_str, copy_user_bytes, copy_to_user_bytes,
  is_valid_user_ptr) take u64 throughout; pointer derefs go via
  `ptr as usize as *mut/*const T` so they stay correct in both 32- and
  64-bit address ranges.
- match arms in dispatch_inner regenerated from handler signatures so
  scalar args get an explicit `as u32` cast and pointer args flow
  through as u64.

Compositor now boots past framebuffer mapping, sys_time, and thread
creation; remaining boot regressions (userspace ALLOC ERROR after
create_window) are tracked separately.

+++ kernel/src/syscall/handlers/debug.rs
@@ -115,7 +115,7 @@ pub fn sys_debug_get_regs(tid: u32, buf_lo: u32, buf_hi_and_size: u32) -> u32 {
 /// Overwrite the target thread's CpuContext from a user buffer.
 ///
 /// Returns 0 on success, u32::MAX on error.
-pub fn sys_debug_set_regs(tid: u32, buf_ptr: u32, size: u32) -> u32 {
+pub fn sys_debug_set_regs(tid: u32, buf_ptr: u64, size: u32) -> u32 {
     let caller_tid = scheduler::current_tid();
     let buf = buf_ptr as u64;
 
@@ -136,7 +136,7 @@ pub fn sys_debug_set_regs(tid: u32, buf_ptr: u32, size: u32) -> u32 {
 /// Uses CR3-switch pattern for cross-process memory access.
 ///
 /// Returns number of bytes read, or u32::MAX on error.
-pub fn sys_debug_read_mem(tid: u32, addr: u32, size: u32, buf_ptr: u32) -> u32 {
+pub fn sys_debug_read_mem(tid: u32, addr: u32, size: u32, buf_ptr: u64) -> u32 {
     let caller_tid = scheduler::current_tid();
     let target_addr = addr as u64;
     let buf = buf_ptr as u64;
@@ -156,7 +156,7 @@ pub fn sys_debug_read_mem(tid: u32, addr: u32, size: u32, buf_ptr: u32) -> u32 {
 /// Write up to 256 bytes into the target thread's address space.
 ///
 /// Returns number of bytes written, or u32::MAX on error.
-pub fn sys_debug_write_mem(tid: u32, addr: u32, size: u32, buf_ptr: u32) -> u32 {
+pub fn sys_debug_write_mem(tid: u32, addr: u32, size: u32, buf_ptr: u64) -> u32 {
     let caller_tid = scheduler::current_tid();
     let target_addr = addr as u64;
     let buf = buf_ptr as u64;
@@ -217,7 +217,7 @@ pub fn sys_debug_single_step(tid: u32) -> u32 {
 ///
 /// Each entry: (start_addr: u64, end_addr: u64, flags: u64) = 24 bytes.
 /// Returns number of regions written, or u32::MAX on error.
-pub fn sys_debug_get_mem_map(tid: u32, buf_ptr: u32, size: u32) -> u32 {
+pub fn sys_debug_get_mem_map(tid: u32, buf_ptr: u64, size: u32) -> u32 {
     let caller_tid = scheduler::current_tid();
     let buf = buf_ptr as u64;
 
@@ -239,7 +239,7 @@ pub fn sys_debug_get_mem_map(tid: u32, buf_ptr: u32, size: u32) -> u32 {
 ///
 /// Returns the event type (1=breakpoint, 2=single_step, 3=exit),
 /// or 0 if no event pending, or u32::MAX on error.
-pub fn sys_debug_wait_event(tid: u32, buf_ptr: u32, size: u32) -> u32 {
+pub fn sys_debug_wait_event(tid: u32, buf_ptr: u64, size: u32) -> u32 {
     let caller_tid = scheduler::current_tid();
     let buf = buf_ptr as u64;
 
@@ -258,7 +258,7 @@ pub fn sys_debug_wait_event(tid: u32, buf_ptr: u32, size: u32) -> u32 {
 /// fd_count, brk, mmap_next, last_cpu, rip, rsp, cr3.
 ///
 /// Returns number of bytes written, or u32::MAX on error.
-pub fn sys_thread_info_ex(tid: u32, buf_ptr: u32, size: u32) -> u32 {
+pub fn sys_thread_info_ex(tid: u32, buf_ptr: u64, size: u32) -> u32 {
     let buf = buf_ptr as u64;
 
     if !is_valid_user_ptr(buf, size as u64) {
